vulnerable-app/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0

